Refactor Operator.java and ExpressionChar.java
It is unnecessary to have ExpressionChar hold the char values of
operators when the Operator enum can do that without an intermediary.

This commit implements a getRawCharList() method in which the
characters of all operators are returned in an ArrayList.

Also, a precedence value was added to all Operators for infix to postfix
notation conversion.