Add infix to postfix notation algorithm
This algorithm is performed before the tokens are parsed, since the
parenthesis need to be taken care of, instead of having an unnecessary
"parenthesis node" that won't be used for calculation. However,
performing this algorithm after the fact should be tested for
performance/readibility.

This commit currently implements this algorithm "jankily". It works for
now, but future commits should improve the readibility.

Along this algorithm, the following changes have been made:
- Add left/right parenthesis enums to ExpressionChar.java
- Change getRawCharList() to a HashMap