Add ExpressionTree.java
This commit currently implements a parsing tree that handles basic
arithmetic expressions based on the given postfix tokens from
ExpressionParser.

There are still a few bugs related to the parsing of spaces and
parentheses. For example, if there are no spaces within the parsing
string, the parser will interpret two numbers separated by an operator
as one. Additionally, negative values appear to be broken again.