Add ability to define variables within the parser
Up till now, I haven't had a way to define variables within variable
expressions. The parser is now able to parse declared variable values
after the delimiter " ~~ ". This delimiter is temporary, probably,
because I don't really know what else to use. Originally, I wanted the
delimiter to be " || ", but this behaved as an OR operation for the
.split() method.

I have also changed the HashMap key type to String instead of Character
because I plan on using multicharacter variables for calculus variables
such as the range of the integral or the location for the derivative.