Add calculus parameter input into parser
This commit adds the ability to add calculus "parameters" such as which
variable the derivative "derives over" and the range of values to
integrate over.

As of now, most of the code is properly functioning, but I've
encountered an interesting bug. The Hashmap which contains the variables
contain different character arrays than normal. Normal strings have a
standard buffer of 10 characters in its internal array, while the ones
in the HashMap somehow contain only 1. This creates different hash
values, which prevent the calculus node from retrieving the variable.